Relax check_listening_on_http check
If a service does two 302 redirect from the bare domain like:

    http://www.servicename.service.gov.uk/ ->
      https://www.servicename.service.gov.uk/ ->
        https://www.gov.uk/service-slug

Then currently the test will fail as it expects only the first redirect.
Unfortunately with 302s, it follows all the redirects in the chain, so
the resulting location is not simply
https://www.servicename.service.gov.uk/

This rewrites the test to connect to the service domain and simply
expect the eventual location to have a scheme of 'https' which will
match wherever that redirect points to as long as it's secure.