Ignore cookie headings without domain=
Cookies without the domain explicitly set assume the current domain
name:

http://en.wikipedia.org/wiki/HTTP_cookie#Domain_and_Path

This would be bad for a top-level domain (eg `gov.uk`), but is valid for
the typical service subdomain as it's only scoped to their `www.`
domain. Given this, we should correctly pass `set-cookie` headers that
don't contain an explicit domain.